+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 0.12 - (Luxor)
 ### Changed
 - Alerts can now be dismissed
+- Date Picker library
 
 ## 0.11 - Kilcrea
 ### Added